--- src/components/cards/CardSegmentNoBorder.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import { CardSegment } from 'lattice-ui-kit';
-
-const CardSegmentNoBorder = styled(CardSegment)`
-  border: none !important;
-`;
-
-export default CardSegmentNoBorder;
