+++ src/components/cards/index.js
@@ -2,6 +2,6 @@
  * @flow
  */
 
-export { default as CardSegmentNoBorder } from './CardSegmentNoBorder';
 export { default as EntitySetSearchResultCard } from './EntitySetSearchResultCard';
 export { default as ErrorCardSegment } from './ErrorCardSegment';
+export { default as TableCardSegment } from './TableCardSegment';
